update: PopulationComposition| mock生成部分の関数作成
利用する際、jest.mock(query.tsへの相対パス) をテストファイルに含める必要あり
